+++ HISTORY.md
@@ -1,6 +1,14 @@
 # History
 ----
 
+## 0.9.0 / 2015-01-18
+
+- support force, force to revalidate.
+
+```
+this.props.validateFields(['xx'], {force: true}).
+```
+
 ## 0.8.0 / 2015-01-13
 
 - support validate/validateFirst option for getFieldProps